requirements.txt
google-cloud
 
google-cloud-bigquery